requirements.txt
streamlit
easyocr
pandas
opencv-python-headless
gspread
oauth2client
Pillow
numpy